Update parametric-speed-flight.cc
In method UpdateDistance for the analytic polynomial integration algorithm was used (order - i +1) but order is instanciated equal to m_speedParams.size() so it's already one more than the real order of the polynomial so (order - i) is correct. Moreover the creation of a new vector to store the modified coefficients was not necessary and a direct multiplication by (order - i) is more efficient.

Also declaring the double y without initialization was yielding some strange behaviours in the Newton-Raphson method (in version 31+) so now it is initialized to 0.