TcpStubClientApp. report Node ID in logs
This facilitates the tracking of SendPacket in debug logs.
Simulator time is not needed anymore as it is correctly reported by ns3
logging facility (see scenario configuration helper on how to setup
this facility to report time).

Caveat: unfortunately ns3's logging facility should also report the node
ID automatically, but it can't determine the reference node when the
callback is executed.